# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Book Catalog API Documentation",
        "description": "Laravel Book Catalog API Documentation",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get list of authors",
                "description": "Returns list of authors",
                "operationId": "getAuthorsList",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {}
                ]
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Store new author",
                "description": "Returns author data",
                "operationId": "storeAuthor",
                "requestBody": {
                    "description": "Data packet for Test",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAuthorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security": []
                    }
                ]
            }
        },
        "/authors/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get author information",
                "description": "Returns author data",
                "operationId": "getAuthorById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Author id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {}
                ]
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Delete author",
                "description": "Delete author",
                "operationId": "deleteAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Author id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Not found"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Authors"
                ],
                "summary": "Edit author",
                "description": "Returns edited author data",
                "operationId": "editAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Author id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data packet for Test",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreAuthorRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key_security": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreAuthorRequest": {
                "title": "StoreAuthorRequest",
                "description": "Store author request body data",
                "required": [
                    "last_name"
                ],
                "properties": {
                    "first_name": {
                        "title": "first_name",
                        "description": "Author`s name",
                        "type": "string",
                        "items": {
                            "type": "string",
                            "items": {}
                        },
                        "example": "Bob"
                    },
                    "last_name": {
                        "title": "last_name",
                        "description": "Author`s surname",
                        "type": "string",
                        "items": {
                            "type": "string",
                            "items": {}
                        },
                        "example": "Marley"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "api_key_security": {
                "type": "apiKey",
                "description": " Enter api_token in format < Bearer your_token>",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "api_key_security": []
        }
    ]
}